Move admin program views into views.admin.programs package.
