+++ universal-application-tool-0.0.1/app/views/admin/AdminLayout.java
@@ -18,7 +18,7 @@ public class AdminLayout extends BaseHtmlLayout {
   }
 
   /** Renders mainDomContents within the main tag, in the context of the admin layout. */
-  protected Content render(DomContent... mainDomContents) {
+  public Content render(DomContent... mainDomContents) {
     return htmlContent(head(tailwindStyles()), body(main(mainDomContents)));
   }
 }
