+++ universal-application-tool-0.0.1/app/views/admin/programs/ProgramBlockEditView.java
@@ -1,4 +1,4 @@
-package views.admin;
+package views.admin.programs;
 
 import static j2html.TagCreator.a;
 import static j2html.TagCreator.div;
@@ -9,6 +9,7 @@ import play.twirl.api.Content;
 import services.program.BlockDefinition;
 import services.program.ProgramDefinition;
 import views.BaseHtmlView;
+import views.admin.AdminLayout;
 
 public class ProgramBlockEditView extends BaseHtmlView {
