+++ universal-application-tool-0.0.1/app/views/admin/programs/ProgramEditView.java
@@ -1,4 +1,4 @@
-package views.admin;
+package views.admin.programs;
 
 import static j2html.TagCreator.a;
 import static j2html.TagCreator.body;
@@ -11,6 +11,7 @@ import play.mvc.Http.Request;
 import play.twirl.api.Content;
 import services.program.ProgramDefinition;
 import views.BaseHtmlView;
+import views.admin.AdminLayout;
 
 public class ProgramEditView extends BaseHtmlView {
   private final AdminLayout layout;
